rebase: store orig_head as a commit
Using a struct commit rather than a struct oid to hold orig_head means
that we error out straight away if branch being rebased does not point
to a commit. It also simplifies the code than handles finding the
merge base and fork point as it not longer has to convert from an oid
to a commit.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>